requirements.txt
python-telegram-bot==22.0
yadisk==3.2.0
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.0
pydub==0.25.1 